src/requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
onetimepass==1.0.1
PyMySQL==1.0.2
six==1.16.0
Werkzeug==2.2.2
zipp==3.11.0
python-dateutil==2.8.2